Crate re_chunk

A chunk of Rerun data, encoded using Arrow. Used for logging, transport, storage and compute.

Feature flags

serde — Enable (de)serialization using serde.

arrow — Enable conversion to and from arrow-rs types

Re-exports

pub use arrow2::array::Array as Arrow2Array;

pub use re_log_types::EntityPath;

pub use re_log_types::TimeInt;

pub use re_log_types::TimePoint;

pub use re_log_types::Timeline;

pub use re_log_types::TimelineName;

pub use re_types_core::ComponentName;

Modules

arrow 🔒

batcher 🔒

builder 🔒

chunk 🔒

external

helpers 🔒

id 🔒

iter 🔒

latest_at 🔒

merge 🔒

range 🔒

shuffle 🔒

slice 🔒

transport 🔒

util

Structs

Chunk

Dense arrow-based storage of N rows of multi-component multi-temporal data for a specific entity.

ChunkBatcher

Implements an asynchronous batcher that coalesces PendingRows into Chunks based upon the thresholds defined in the associated ChunkBatcherConfig.

ChunkBatcherConfig

Defines the different thresholds of the associated ChunkBatcher.

ChunkBuilder

Helper to incrementally build a Chunk.

ChunkComponentIter

The actual iterator implementation for Chunk::iter_component.

ChunkComponentIterItem

The underlying item type for ChunkComponentIter.

ChunkId

A unique ID for a crate::Chunk.

ChunkIndicesIter

LatestAtQuery

A query at a given time, for a given timeline.

PendingRow

A single row’s worth of data (i.e. a single log call).

RangeQuery

A query over a time range, for a given timeline.

RangeQueryOptions

RowId

A unique ID for a row’s worth of data within a chunk.

TimeColumn

TimeColumnBuilder

Helper to incrementally build a TimeColumn.

TransportChunk

A Chunk that is ready for transport. Obtained by calling Chunk::to_transport.

UnitChunkShared

A ChunkShared that is guaranteed to always contain a single row’s worth of data.

Enums

ChunkBatcherError

Errors that can occur when creating/manipulating a ChunkBatcher.

ChunkError

Errors that can occur when creating/manipulating a Chunks, directly or indirectly through the use of a crate::ChunkBatcher.

Type Aliases

ChunkBatcherResult

ChunkResult

ChunkShared

A simple type alias for an Arc<Chunk>.