Crate re_chunk

A chunk of Rerun data, encoded using Arrow. Used for logging, transport, storage and compute.

Feature flags

No documented features in Cargo.toml

Modules

batcher 🔒

chunk 🔒

external

shuffle 🔒

transport 🔒

util 🔒

Structs

Chunk

Dense arrow-based storage of N rows of multi-component multi-temporal data for a specific entity.

ChunkBatcher

Implements an asynchronous batcher that coalesces PendingRows into Chunks based upon the thresholds defined in the associated ChunkBatcherConfig.

ChunkBatcherConfig

Defines the different thresholds of the associated ChunkBatcher.

ChunkTimeline

PendingRow

A single row’s worth of data (i.e. a single log call).

TransportChunk

A Chunk that is ready for transport. Obtained by calling Chunk::to_transport.

Enums

ChunkBatcherError

Errors that can occur when creating/manipulating a ChunkBatcher.

ChunkError

Errors that can occur when creating/manipulating a Chunks, directly or indirectly through the use of a crate::ChunkBatcher.

Functions

arrays_to_list_array

Create a sparse list-array out of an array of arrays.

Type Aliases

ChunkBatcherResult

ChunkId

Unique identifier for a Chunk, using a [re_tuid::Tuid].

ChunkResult