Crate re_chunk

A chunk of Rerun data, encoded using Arrow. Used for logging, transport, storage and compute.

Feature flags

serde — Enable (de)serialization using serde.

Re-exports

pub use arrow::array::Array as ArrowArray;

pub use re_log_types::EntityPath;

pub use re_log_types::TimeInt;

pub use re_log_types::TimePoint;

pub use re_log_types::Timeline;

pub use re_log_types::TimelineName;

pub use re_types_core::ArchetypeFieldName;

pub use re_types_core::ArchetypeName;

pub use re_types_core::ChunkId;

pub use re_types_core::ComponentName;

pub use re_types_core::RowId;

Modules

batcher 🔒

builder 🔒

chunk 🔒

external

helpers 🔒

iter 🔒

latest_at 🔒

merge 🔒

migration 🔒

range 🔒

shuffle 🔒

slice 🔒

transport 🔒

Structs

Chunk

Dense arrow-based storage of N rows of multi-component multi-temporal data for a specific entity.

ChunkBatcher

Implements an asynchronous batcher that coalesces PendingRows into Chunks based upon the thresholds defined in the associated ChunkBatcherConfig.

ChunkBatcherConfig

Defines the different thresholds of the associated ChunkBatcher.

ChunkBuilder

Helper to incrementally build a Chunk.

ChunkComponentIter

The actual iterator implementation for Chunk::iter_component.

ChunkComponentIterItem

The underlying item type for ChunkComponentIter.

ChunkComponents

ChunkIndicesIter

LatestAtQuery

A query at a given time, for a given timeline.

PendingRow

A single row’s worth of data (i.e. a single log call).

RangeQuery

A query over a time range, for a given timeline.

RangeQueryOptions

TimeColumn

TimeColumnBuilder

Helper to incrementally build a TimeColumn.

UnitChunkShared

A ChunkShared that is guaranteed to always contain a single row’s worth of data.

Enums

ChunkBatcherError

Errors that can occur when creating/manipulating a ChunkBatcher.

ChunkError

Errors that can occur when creating/manipulating a Chunks, directly or indirectly through the use of a crate::ChunkBatcher.

TimeColumnError

Errors when deserializing/parsing/reading a column of time data.

Traits

ChunkComponentSlicer

A ChunkComponentSlicer knows how to efficiently slice component batches out of a Chunk column.

Type Aliases

ChunkBatcherResult

ChunkResult

ChunkShared

A simple type alias for an Arc<Chunk>.